+++ pkg/commands/ngwaf/workspace/timeseries/root.go
@@ -1,0 +1,29 @@
+package timeseries
+
+import (
+	"io"
+
+	"github.com/fastly/cli/pkg/argparser"
+	"github.com/fastly/cli/pkg/global"
+)
+
+// RootCommand is the parent command for all subcommands in this package.
+type RootCommand struct {
+	argparser.Base
+}
+
+// CommandName is the string to be used to invoke this command.
+const CommandName = "time-series"
+
+// NewRootCommand returns a new command registered in the parent.
+func NewRootCommand(parent argparser.Registerer, g *global.Data) *RootCommand {
+	var c RootCommand
+	c.Globals = g
+	c.CmdClause = parent.Command(CommandName, "Manage workspace time series metrics")
+	return &c
+}
+
+// Exec implements the command interface.
+func (c *RootCommand) Exec(_ io.Reader, _ io.Writer) error {
+	panic("unreachable")
+}